Embed the builder classes and simplify
