+++ sdk/java/pulumi/src/main/java/com/pulumi/deployment/InvokeOptions.java
@@ -38,10 +38,7 @@ public class InvokeOptions {
         @Nullable Resource parent, 
         @Nullable ProviderResource provider, 
         @Nullable String version) {
-        this.parent = parent;
-        this.provider = provider;
-        this.version = version;
-        this.pluginDownloadURL = null;
+        this(parent, provider, version, null);
     }
 
     public InvokeOptions(
@@ -88,8 +85,56 @@ public class InvokeOptions {
         return Optional.ofNullable(pluginDownloadURL);
     }
 
-    public static InvokeOptionsBuilder builder() {
-        return new InvokeOptionsBuilder();
+    public static Builder builder() {
+        return new Builder();
+    }
+
+    public static final class Builder {
+        private @Nullable Resource parent;
+        private @Nullable ProviderResource provider;
+        private @Nullable String version;
+        private @Nullable String pluginDownloadURL;
+
+        /**
+         * An optional parent resource to which this invoke belongs.
+         */
+        public Builder parent(Resource parent) {
+            this.parent = parent;
+            return this;
+        }
+    
+        /**
+         * An optional provider to use for this invoke. If no provider is
+         * supplied, the default provider for the invoke package will be used.
+         */
+        public Builder provider(ProviderResource provider) {
+            this.provider = provider;
+            return this;
+        }
+
+        /**
+         * An optional version, corresponding to the version of the provider plugin that should be
+         * used when operating on this invoke. This version overrides the version information
+         * inferred from the current package and should rarely be used.
+         */
+        public Builder version(String version) {
+            this.version = version;
+            return this;
+        }
+
+        /**
+         * An optional URL, corresponding to the url from which the provider plugin that should be
+         * used when operating on this invoke is downloaded from. This URL overrides the download URL
+         * inferred from the current package and should rarely be used.
+         */
+        public Builder pluginDownloadURL(String pluginDownloadURL) {
+            this.pluginDownloadURL = pluginDownloadURL;
+            return this;
+        }
+    
+        public InvokeOptions build() {
+            return new InvokeOptions(parent, provider, version, pluginDownloadURL);
+        }
     }
 
     @InternalUse
