+++ sdk/java/pulumi/src/main/java/com/pulumi/deployment/InvokeOutputOptions.java
@@ -25,10 +25,6 @@ public class InvokeOutputOptions extends InvokeOptions {
 
     private final List<Resource> dependsOn;
 
-    public InvokeOutputOptions() {
-        this(null, null, null, null, null);
-    }
-
     public InvokeOutputOptions(
         @Nullable Resource parent, 
         @Nullable ProviderResource provider, 
@@ -44,8 +40,11 @@ public class InvokeOutputOptions extends InvokeOptions {
         @Nullable ProviderResource provider, 
         @Nullable String version,
         @Nullable List<Resource> dependsOn) {
-        super(parent, provider, version, null);
-        this.dependsOn = dependsOn;
+        this(parent, provider, version, null, dependsOn);
+    }
+
+    public InvokeOutputOptions() {
+        this(null, null, null, null, null);
     }
 
     /**
@@ -54,6 +53,68 @@ public class InvokeOutputOptions extends InvokeOptions {
      */
     public List<Resource> getDependsOn() {
         return this.dependsOn == null ? List.of() : new ArrayList<>(this.dependsOn);
+    }
+
+    public static final class Builder {
+        private @Nullable Resource parent;
+        private @Nullable ProviderResource provider;
+        private @Nullable String version;
+        private @Nullable String pluginDownloadURL;
+        private @Nullable List<Resource> dependsOn;
+    
+        /**
+         * An optional parent resource to which this invoke belongs.
+         */
+        public Builder parent(Resource parent) {
+            this.parent = parent;
+            return this;
+        }
+    
+        /**
+         * An optional provider to use for this invoke. If no provider is
+         * supplied, the default provider for the invoke package will be used.
+         */
+        public Builder provider(ProviderResource provider) {
+            this.provider = provider;
+            return this;
+        }
+    
+        /**
+         * An optional version, corresponding to the version of the provider plugin that should be
+         * used when operating on this invoke. This version overrides the version information
+         * inferred from the current package and should rarely be used.
+         */
+        public Builder version(String version) {
+            this.version = version;
+            return this;
+        }
+    
+        /**
+         * An optional URL, corresponding to the url from which the provider plugin that should be
+         * used when operating on this invoke is downloaded from. This URL overrides the download URL
+         * inferred from the current package and should rarely be used.
+         */
+        public Builder pluginDownloadURL(String pluginDownloadURL) {
+            this.pluginDownloadURL = pluginDownloadURL;
+            return this;
+        }
+    
+        /**
+         *  Additional explicit dependencies on other resources.
+         */
+        public Builder dependsOn(Resource... dependsOn) {
+            if (this.dependsOn == null) {
+                this.dependsOn = new ArrayList<Resource>();
+            }
+            for (var r : dependsOn) {
+                this.dependsOn.add(r);
+            }
+            return this;
+        }
+    
+        public InvokeOutputOptions build() {
+            return new InvokeOutputOptions(parent, provider, version, pluginDownloadURL, dependsOn);
+        }
     }
 
     @InternalUse
